// pinturillo-api/src/entities/enums.ts
// ** Enums **
export enum EVENTS_SOCKET_CLIENT {
  CONNECT = "custom connect",
  DISCONNECT = "disconnect",
  JOIN_GAME = "join game",
  JOIN_LOBBY = "join lobby",
  JOIN_ROOM_LOBBY = "join room lobby",
  LEAVE_LOBBY = "leave lobby",
  SEND_MESSAGE_LOBBY = "send message lobby",
  CREATE_ROOM = "create room",
  LOGIN_PRIVATE_ROOM = "login private room",
  START_GAME = "start game",
  WORD_SELECTED_GAME = "word selected game",
  CANVAS_IMAGE_GAME = "canvas image game",
  CANVAS_CLEAR_GAME = "canvas clear game",
  COUTNDOWN_GAME = "countdown game",
  SEND_MESSAGE_GAME = "send message game",
  NEW_PAINTER_GAME = "new painter game",
  NEXT_ROUND_GAME = "next round game",
  FINISH_GAME = "finish game",
}

export enum EVENTS_SOCKET_SERVER {
  CONNECT = "custom connect", 
  DISCONNECT = "custom disconnect", 
  UPDATE_LOBBY = "update lobby", 
  UPDATE_ROOM_LOBBY = "update room lobby", 
  UPDATE_GAME = "update game", 
  SEND_MESSAGE_LOBBY = "send message lobby", 
  LOGIN_PRIVATE_ROOM = "login private room", 
  CREATE_ROOM = "create room", 
  START_GAME = "start game", 
  CANVAS_IMAGE_GAME = "canvas image game", 
  CANVAS_CLEAR_GAME = "canvas clear game", 
  SEND_MESSAGE_GAME = "send message game",
  NEW_PAINTER_GAME = "new painter game",
  NEXT_ROUND_GAME = "next round game",
  FINISH_GAME = "finish game",
}